move cell out of transform dir
