+++ CMakeLists.txt
@@ -76,7 +76,7 @@ target_include_directories(test_sl PRIVATE common)
 target_link_libraries(test_sl common)
 add_test(NAME sl COMMAND test_sl)
 
-add_executable(test_cell test/cell.c src/crmap.c src/transform/cell.c)
+add_executable(test_cell test/cell.c src/crmap.c src/cell.c)
 target_include_directories(test_cell PRIVATE common src)
 target_link_libraries(test_cell common cjson)
 add_test(NAME cell COMMAND test_cell)
